object/requirement.txt
numpy
opencv-python
kafka-python
ultralytics